Actualitzada branca amb els canvis r1895:1943  del trunk
git-svn-id: http://devel.cpl.upc.edu/recursos/svn/smeagol/branches/tiquet_370@1944 4ab63a03-c2d3-4037-8559-9a21b594da05

+++ lib/V2/Server/Controller/Booking.pm
@@ -382,7 +382,7 @@ sub default_POST {
     my ( $self, $c, $res, $id, $module, $id_module ) = @_;
     my $req = $c->request;
 
-	if(($module eq 'tag') && ($id_module)){
+	if((defined $module) && ($module eq 'tag') && ($id_module)){
 		$c->detach( 'post_relation_tag_booking');
 	}
 
@@ -409,9 +409,6 @@ sub default_POST {
         );
     }
     
-    my @tags = split( ',', $req->parameters->{tags} )
-        if defined $req->parameters->{tags};
-
 #dtstart and dtend are parsed in case that some needed parameters to build the recurrence of the
 #booking aren't provided
     $dtstart  = ParseDate($dtstart);
@@ -467,7 +464,6 @@ sub default_POST {
                 duration    => $new_booking->duration,
                 by_minute   => $new_booking->by_minute,
                 by_hour     => $new_booking->by_hour,
-                tags        => \@tags,
             };
         }
 
@@ -500,7 +496,6 @@ sub default_POST {
                 by_minute   => $new_booking->by_minute,
                 by_hour     => $new_booking->by_hour,
                 by_day      => $new_booking->by_day,
-                tags        => \@tags,
             };
 
         }
@@ -535,7 +530,6 @@ sub default_POST {
                 by_hour      => $new_booking->by_hour,
                 by_month     => $new_booking->by_month,
                 by_day_month => $new_booking->by_day_month,
-                tags        => \@tags,
             };
         }
 
@@ -571,7 +565,6 @@ sub default_POST {
                 by_day       => $new_booking->by_day,
                 by_month     => $new_booking->by_month,
                 by_day_month => $new_booking->by_day_month,
-                tags        => \@tags,
             };
 
         }
@@ -582,11 +575,9 @@ sub default_POST {
         $c->stash->{new_exception} = $exception;
     }
 
-    $c->visit( '/check/check_overlap', [] );
+    $c->forward( '/check/check_overlap', [] );
     my @message;
-    
-    my $id_tag; my $tags;
-    
+       
     if ( $c->stash->{booking_ok} == 1 ) {
 	 
         if (   $c->stash->{overlap} == 1
@@ -594,14 +585,11 @@ sub default_POST {
             or $c->stash->{too_long} == 1 )
         {
             if ( $c->stash->{empty} == 1 ) {
-                @message = { message => "Bad Request", };
+                #TODO: message: parametres estan malament
                 $c->response->status(400);
             }
             else {
-                @message
-                    = { message =>
-                        "Error: The booking you tried to create overlaps with another booking or with itself",
-                    };
+                #TODO: message: Booking amb resource ocupat
                 $c->response->status(409);
             }
 
@@ -613,36 +601,20 @@ sub default_POST {
         else {
 
 	     $new_booking->insert;
-	    
-	     foreach (@tags) {
-		  $id_tag = $_;
-		  
-		  $tags = $c->model('DB::TTag')->find( { id => $id_tag } );
-		  
-		  if ($tags) {
-		       my $tag_booking = $c->model('DB::TTagBooking')->find_or_new();
-		       $tag_booking->id_tag($id_tag);
-		       $tag_booking->id_booking( $new_booking->id );
-		       $tag_booking->insert;
-		  }
-		  else {
-		       $c->detach( '/bad_request', [] );
-		  }
-	     }
-
-            $c->stash->{content} = $booking;
+	   
+	    #TODO: booking creat correctament
+            $c->stash->{content} = \@message;
             $c->stash->{booking} = $booking;
             $c->response->status(201);
 
             #$c->stash->{template} = 'booking/get_booking.tt';
-            $c->forward( 'get_booking', [ $new_booking->id ] );
+            #$c->forward( 'get_booking', [ $new_booking->id ] );
+	    
 
         }
     }
     else {
-        my @message
-            = { message => "Error: Check if the event or the resource exist",
-            };
+		#TODO: message: parametres estan malament   
         $c->stash->{content} = \@message;
         $c->response->status(400);
         $c->stash->{error}
@@ -716,9 +688,6 @@ sub put_booking : Private {
     $c->stash->{id_event}    = $id_event;
     $c->stash->{id_resource} = $id_resource;
 
-    my @tags = split( ',', $req->parameters->{tags} )
-        if defined $req->parameters->{tags};
-    
     #Do the resource and the event exist?
     $c->visit( '/check/check_booking', [] );
 
@@ -755,7 +724,6 @@ sub put_booking : Private {
                 duration    => $booking->duration,
                 by_minute   => $booking->by_minute,
                 by_hour     => $booking->by_hour,
-                tags        => \@tags,
             };
         }
 
@@ -787,7 +755,6 @@ sub put_booking : Private {
                 by_minute   => $booking->by_minute,
                 by_hour     => $booking->by_hour,
                 by_day      => $booking->by_day,
-                tags        => \@tags,
             };
 
         }
@@ -822,7 +789,6 @@ sub put_booking : Private {
                 by_hour      => $booking->by_hour,
                 by_month     => $booking->by_month,
                 by_day_month => $booking->by_day_month,
-                tags        => \@tags,
             };
         }
 
@@ -858,7 +824,6 @@ sub put_booking : Private {
                 by_day       => $booking->by_day,
                 by_month     => $booking->by_month,
                 by_day_month => $booking->by_day_month,
-                tags        => \@tags,
             };
 
         }
@@ -871,7 +836,7 @@ sub put_booking : Private {
     $c->stash->{PUT}         = 1;
     $c->visit( '/check/check_overlap', [] );
 
-    my $id_tag; my $tags;
+
     
     if ( $c->stash->{booking_ok} == 1 ) {
 
@@ -889,23 +854,6 @@ sub put_booking : Private {
         else {
             $booking->update;
 	    
-	    foreach (@tags) {
-		 $id_tag = $_;
-		 
-		 $tags = $c->model('DB::TTag')->find( { id => $id_tag } );
-		 
-		 if ($tags) {
-		      my $tag_booking = $c->model('DB::TTagBooking')->find_or_new();
-		      $tag_booking->id_tag($id_tag);
-		      $tag_booking->id_booking( $booking->id );
-		      $tag_booking->insert;
-		 }
-		   else {
-			$c->detach( '/bad_request', [] );
-		   }
-	    }
-	    
-
             $c->stash->{content} = $jbooking;
             $c->stash->{booking} = $jbooking;
             $c->response->status(201);
@@ -953,7 +901,7 @@ sub default_DELETE {
     my $req = $c->request;
     
 if ($id) {
-    if(($module eq 'tag') && ($id_module)){
+    if((defined $module) && ($module eq 'tag') && ($id_module)){
         $c->detach( 'delete_relation_tag_booking', [$id, $id_module]);
     }
     else {
@@ -963,6 +911,7 @@ if ($id) {
 	    $booking_aux->delete;
 	    #TODO: message: Resource esborrat amb èxit.
 	    $c->stash->{content}  = \@message;
+		$c->response->status(200);
         }
         else {  
 	    #TODO: message: Resource no trobat.
@@ -1325,3 +1274,4 @@ it under the same terms as Perl itself.
 =cut
 
 1;
+
